Feature/factory reset (#186)
- Enable factory reset via file in home directory.
- AIIDALAB_FACTORY_RESET environment variable still takes precedence.
- The reset mode is implemented as bit mask to allow fine-grained
  control over the reset.

+++ my_init.d/factory_reset.sh
@@ -1,34 +1,43 @@
-#!/bin/bash -e
-
-# Debugging.
-set -x
+#!/bin/bash -eu
 
 # Environment.
 export SHELL=/bin/bash
 
-# Performing factory reset of your AiiDAlab environment:
-# 0 - No reset (noop).
-# 1 - Remove locally installed software and apps (removes ~/apps/ and ~/.local/).
-# 2 - Remove all files and directories within the users home directory.
+# Bit mask for performing factory reset of the AiiDAlab environment:
+# 0b000 0 - No reset (noop).
+# 0b001 1 - Remove locally installed software and apps (removes ~/apps/ and ~/.local/).
+# 0b010 2 - Remove all files and directories within the users home directory.
 
+# If the ~/AIIDALAB_FACTORY_RESET file exists, parse it (and remove it).
+if [ -e "/home/${SYSTEM_USER}/AIIDALAB_FACTORY_RESET" ]; then
+  RESET_MODE="`cat /home/${SYSTEM_USER}/AIIDALAB_FACTORY_RESET`"
+  rm -f "/home/${SYSTEM_USER}/AIIDALAB_FACTORY_RESET"  # Remove file after parsing.
+fi
 
-case "${AIIDALAB_FACTORY_RESET}" in
+# If the AIIDALAB_FACTORY_RESET environment variable is set, it takes preference (default mode=0):
+RESET_MODE="${AIIDALAB_FACTORY_RESET:-${RESET_MODE:-0}}"
 
-  0)
-    exit 0
-    ;;
+# Warn about unknown mode.
+if (( ${RESET_MODE} > 0x3 )); then
+  echo "factory reset: WARNING: UNKNOWN RESET MODE (${RESET_MODE}) CONTAINS UNKNOWN BITS."
+fi
 
-  1)
-    rm -rf "/home/${SYSTEM_USER}/apps"
-    rm -rf "/home/${SYSTEM_USER}/.local"
-    ;;
+# Perform the individual reset functions:
 
-  2)
-    find /home/${SYSTEM_USER}/ -mindepth 1 -delete
-    ;;
+# mode==0 -> noop
+if (( ${RESET_MODE} == 0)); then  # noop
+    echo "factory reset: No reset (mode=0)."
+fi
 
-  *)
-    echo "Unknown factory reset mode: ${AIIDALAB_FACTORY_RESET}"
-    exit 1
-    ;;
-esac
+# mode & 001 -> delete ~/apps/ and ~/.local/
+if (( (${RESET_MODE} & 0x1) == 0x1)); then
+  echo "factory reset: Remove apps (~/apps/) and local software installation (~/.local/)."
+  rm -rf "/home/${SYSTEM_USER}/apps"
+  rm -rf "/home/${SYSTEM_USER}/.local"
+fi
+
+# mode & 010 -> delete all home directory contents
+if (( (${RESET_MODE} & 0x2) == 0x2)); then
+  echo "factory reset: Remove user home directory contents."
+  find "/home/${SYSTEM_USER}/" -mindepth 1 -delete
+fi
